# Improve performance of equals and equalsUnchecked