@fmeiser 2016-09-30
@martint Another small improvement in performance.