@martint 2016-09-30
Merged, thanks!